backend/requirements.txt
fastapi
uvicorn
requests
faiss-cpu
numpy
pydantic
aiofiles